+++ packages/core/src/components/user-menu.ts
@@ -4,6 +4,7 @@ export const UserMenu = () => ({
     '&-button': {
       '@apply p-0 bg-transparent hover:bg-transparent': {},
       '@apply border-0': {},
+      '@apply hover:border-0': {},
       '.sk-avatar': {
         height: 'inherit',
         width: 'inherit',