+++ packages/menubar/src/menubar-item.tsx
@@ -1,4 +1,5 @@
-import { DefaultProps, cx, getValidChildren } from '@sk-web-gui/utils';
+import { DefaultProps, cx, getValidChildren, useForkRef } from '@sk-web-gui/utils';
+import { PopupMenu } from '@sk-web-gui/popup-menu';
 import React from 'react';
 import { useMenuBar } from './menubar';
 
@@ -33,42 +34,89 @@ export const MenuBarItem = React.forwardRef<HTMLLIElement, MenuBarItemProps>((pr
   React.useEffect(() => {
     if (mounted) {
       if (active === menuIndex) {
-        menuRef.current && menuRef.current.focus();
+        if (menuRef.current) {
+          walkFocus(menuRef.current.children);
+        }
       }
     } else {
       setMounted(true);
     }
   }, [active]);
 
+  React.useEffect(() => {
+    return () => {
+      setMounted(false);
+    };
+  }, []);
+
+  const setFocus = (element: Element): boolean => {
+    if (element) {
+      if (element.getAttribute('role') === 'menuitem') {
+        (element as HTMLElement).focus();
+        return true;
+      }
+    }
+    return false;
+  };
+
+  const walkFocus = (collection: HTMLCollection): boolean => {
+    let focused = false;
+    const elements = Array.from(collection);
+    for (let index = 0; index < elements.length; index++) {
+      if (!focused) {
+        focused = setFocus(elements[index]);
+        if (!focused && elements[index].children) {
+          focused = walkFocus(elements[index].children);
+        }
+        if (focused) {
+          break;
+        }
+      } else {
+        break;
+      }
+    }
+
+    return focused;
+  };
+
   const handleKeyboard = (event: KeyboardEvent) => {
     if (event.key === 'ArrowLeft') {
       event.preventDefault();
+      event.stopPropagation();
       prev && prev();
     }
     if (event.key === 'ArrowRight') {
       event.preventDefault();
+      event.stopPropagation();
       next && next();
-    }
-    if (event.key === 'Enter') {
-      event.target?.dispatchEvent(new MouseEvent('click'));
     }
   };
 
   const getClonedChild = (child: JSX.Element): React.ReactNode => {
     if (child.type === React.Fragment) {
-      const grandchild = getValidChildren(child.props.children)[0];
-      if (grandchild) {
-        return React.cloneElement(child, { ...child.props, children: getClonedChild(grandchild) });
-      }
+      const grandchildren = getValidChildren(child.props.children).map((child, index) => {
+        return index === 0 ? getClonedChild(child) : child;
+      });
+      return React.cloneElement(child, { ...child.props, children: grandchildren });
+    } else if (child.type === PopupMenu) {
+      const grandchildren = getValidChildren(child.props.children).map((child) => {
+        if (child.type === PopupMenu.Button) {
+          const button = getClonedChild(child);
+          return button;
+        }
+        return child;
+      });
+
+      return React.cloneElement(child, { ...child.props, children: grandchildren });
+    } else {
+      return React.cloneElement(child, {
+        onKeyDown: handleKeyboard,
+        role: 'menuitem',
+        'aria-current': current === menuIndex ? 'page' : undefined,
+        tabIndex: active === menuIndex ? 0 : -1,
+        ...child.props,
+      });
     }
-    return React.cloneElement(child, {
-      onKeyDown: handleKeyboard,
-      ref: menuRef,
-      role: 'menuitem',
-      'aria-current': current === menuIndex ? 'page' : undefined,
-      tabIndex: active === menuIndex ? 0 : -1,
-      ...children.props,
-    });
   };
 
   const getChildWithWrapper = () => {
@@ -80,7 +128,13 @@ export const MenuBarItem = React.forwardRef<HTMLLIElement, MenuBarItemProps>((pr
   };
 
   return (
-    <li data-color={color} ref={ref} className={cx('sk-menubar-item', className)} role="none" {...rest}>
+    <li
+      data-color={color}
+      ref={useForkRef(ref, menuRef)}
+      className={cx('sk-menubar-item', className)}
+      role="none"
+      {...rest}
+    >
       {getChildWithWrapper()}
     </li>
   );