+++ packages/modal/src/dialog/index.ts
@@ -1,7 +1,7 @@
 import React from 'react';
-import DialogComponent, { InternalDialogProps } from './dialog';
-import DialogButtons from './dialog-buttons';
-import DialogContent from './dialog-content';
+import { DialogComponent, InternalDialogProps } from './dialog';
+import { DialogButtons } from './dialog-buttons';
+import { DialogContent } from './dialog-content';
 
 interface DialogProps extends React.ForwardRefExoticComponent<InternalDialogProps> {
   Component: typeof DialogComponent;