+++ packages/popup-menu/src/popup-menu.tsx
@@ -1,77 +1,158 @@
-import { Menu } from '@headlessui/react';
-import { DefaultProps } from '@sk-web-gui/utils';
-import { cx, __DEV__ } from '@sk-web-gui/utils';
+import { __DEV__, useForkRef } from '@sk-web-gui/utils';
 import React from 'react';
+import { useOnClickOutside } from 'usehooks-ts';
 import { PopupMenuButton } from './popup-menu-button';
+import { PopupMenuPanel } from './popup-menu-panel';
+import { GoTo, PopupMenuContext } from './popupmenu-context';
 
-export interface PopupMenuPropsInternal extends React.ComponentPropsWithRef<'button'>, DefaultProps {
+export interface PopupMenuBaseProps {
+  /**
+   * Size of popup. Does not affect button size
+   * @default md
+   */
   size?: 'md' | 'sm';
+  /**
+   * Position of the popup relative to the button.
+   * @default under
+   */
   position?: 'under' | 'over' | 'left' | 'right';
+  /**
+   * Alignment of the popup relative to the button.
+   * @default start
+   */
   align?: 'start' | 'end';
+  /**
+   * Type of popup.
+   * @default menu
+   */
+  type?: 'dialog' | 'menu';
 }
 
-export const PopupMenuComponent = React.forwardRef<HTMLButtonElement, PopupMenuPropsInternal>((props, ref) => {
-  const { className, children, size = 'md', position = 'under', align = 'start', ...rest } = props;
+export interface PopupMenuComponentProps extends PopupMenuBaseProps {
+  /**
+   * PopupMenu.Button & PopupMenu.Panel
+   */
+  children:
+    | [React.ReactComponentElement<typeof PopupMenuButton>, React.ReactComponentElement<typeof PopupMenuPanel>]
+    | React.ReactComponentElement<typeof PopupMenuPanel>;
 
-  const getButton = () => {
-    const button = React.Children.toArray(children).find(
-      (child) =>
-        React.isValidElement(child) && typeof child.type !== 'string' && child?.type?.name === PopupMenuButton.name
-    ) as React.ReactElement;
+  id?: string;
+  /**
+   * Set true to open the menu
+   * @default false
+   */
+  open?: boolean;
+  /**
+   * Returns true or false when menu is opened or closed
+   */
+  onToggleOpen?: (open: boolean) => void;
+}
 
-    if (button) {
-      return React.cloneElement(button, { ...button.props, ref, ...rest });
-    } else {
-      throw new Error('No PopupMenu.Button found');
+export const PopupMenuComponent: React.FC<PopupMenuComponentProps> = (props) => {
+  const [isOpen, setIsOpen] = React.useState<boolean>(false);
+  const [goTo, setGoTo] = React.useState<GoTo | undefined>();
+  const [buttonId, setButtonId] = React.useState<string>();
+  const {
+    children,
+    size = 'md',
+    position = 'under',
+    align = 'start',
+    type = 'dialog',
+    id: _id,
+    open: _open,
+    onToggleOpen,
+  } = props;
+  const buttonRef = React.useRef<HTMLButtonElement>(null);
+  const internalRef = React.useRef<HTMLDivElement>(null);
+  const autoId = React.useId();
+
+  const id = _id || `sk-popup-menu-${autoId}`;
+  const open = () => {
+    setIsOpen(true);
+    onToggleOpen && onToggleOpen(true);
+  };
+
+  const close = (focusButton?: boolean) => {
+    setIsOpen(false);
+    onToggleOpen && onToggleOpen(false);
+    if (focusButton) {
+      buttonRef.current && buttonRef.current.focus();
     }
   };
 
-  const getItems = () => {
-    return React.Children.toArray(children).filter(
-      (child) =>
-        React.isValidElement(child) && typeof child.type !== 'string' && child?.type?.name !== PopupMenuButton.name
-    );
+  React.useEffect(() => {
+    if (_open) {
+      open();
+    } else {
+      close();
+    }
+  }, [_open]);
+
+  useOnClickOutside(internalRef, (event: MouseEvent) => {
+    const target = event.target as Node;
+    const isOutsideButton = buttonRef.current && !buttonRef.current.contains(target);
+    if (isOutsideButton) {
+      close();
+    }
+  });
+
+  const handleOpen = (goTo?: GoTo) => {
+    if (goTo) {
+      setGoTo(goTo);
+    }
+    open();
   };
 
-  return (
-    <Menu>
-      {getButton()}
-      <Menu.Items
-        className={cx('sk-popup-menu-items', className, `sk-popup-menu-items-${size}`)}
-        data-position={position}
-        data-align={align}
-      >
-        {getItems()}
-      </Menu.Items>
-    </Menu>
-  );
-});
-
-export interface PopupMenuItemProps {
-  children: JSX.Element;
-  disabled?: boolean;
-}
-
-//TODO: Add function for expandable menues.
-
-export const PopupMenuItem: React.FC<PopupMenuItemProps> = (props) => {
-  const getItem = (child: JSX.Element, active: boolean) => {
-    const classes = cx(child?.props?.className, active ? 'active' : '');
-    const props = { ...child.props, className: cx('sk-popup-menu-item', classes) };
-    return React.cloneElement(child, props);
+  const context = {
+    size,
+    position,
+    align,
+    isOpen,
+    open: handleOpen,
+    close,
+    goTo,
+    setGoTo,
+    type,
+    id,
+    buttonId,
+    setButtonId,
   };
 
+  const getButton = React.useCallback(() => {
+    const button = React.Children.toArray(children).find(
+      (child) => React.isValidElement(child) && typeof child.type !== 'string' && child?.type === PopupMenuButton
+    ) as React.ReactElement;
+
+    if (button) {
+      return React.cloneElement(button, {
+        ...button.props,
+        ref: button.props.ref ? useForkRef(button.props.ref, buttonRef) : buttonRef,
+      });
+    }
+  }, [children]);
+
+  const getPanel = React.useCallback(() => {
+    const panel = React.Children.toArray(children).find(
+      (child) => React.isValidElement(child) && typeof child.type !== 'string' && child?.type === PopupMenuPanel
+    ) as React.ReactElement;
+    if (panel) {
+      return React.cloneElement(panel, {
+        ...panel.props,
+        ref: panel.props.ref ? useForkRef(panel.props.ref, internalRef) : internalRef,
+      });
+    } else {
+      throw new Error('No PopupMenu.Panel found');
+    }
+  }, [children]);
+
   return (
-    <Menu.Item {...props} disabled={props.disabled}>
-      {({ active }) => getItem(props.children, active)}
-    </Menu.Item>
+    <>
+      <PopupMenuContext.Provider value={context}>
+        {getButton()}
+        {getPanel()}
+      </PopupMenuContext.Provider>
+    </>
   );
-};
-
-export const PopupMenuGroup: React.FC<React.ComponentPropsWithRef<'div'> & DefaultProps> = (props) => {
-  const { className, ...rest } = props;
-
-  return <div role="group" className={cx('sk-popup-menu-items-group', className)} {...rest} />;
 };
 
 if (__DEV__) {