+++ packages/user-menu/stories/user-menu.stories.tsx
@@ -1,10 +1,7 @@
 import React from 'react';
 import { cx } from '@sk-web-gui/utils';
 import { Meta } from '@storybook/react';
-import { Link } from '@sk-web-gui/react';
-import { UserMenu } from '../src';
-import { MenuItemGroup, UserMenuProps } from '../src/user-menu';
-import { Icon } from '@sk-web-gui/icon';
+import { UserMenu, Link, Icon, MenuItemGroup, UserMenuProps } from '@sk-web-gui/react';
 
 export default {
   title: 'Komponenter/UserMenu',