+++ packages/react/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.4](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@1.2.3...@sk-web-gui/react@1.2.4) (2024-02-27)
+
+**Note:** Version bump only for package @sk-web-gui/react
+
 ## [1.2.3](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@1.2.2...@sk-web-gui/react@1.2.3) (2024-02-02)
 
 **Note:** Version bump only for package @sk-web-gui/react